Cleanup

+++ src/ui/Controller/Controller.cpp
@@ -21,14 +21,14 @@ namespace mud
 	{
 		if(!widget.m_events) return;
 
-		KeyEvent* key_down_event = static_cast<KeyEvent*>(widget.m_events->m_events[size_t(DeviceType::Keyboard)][size_t(EventType::Pressed)]);
+		KeyEvent* key_down_event = static_cast<KeyEvent*>(widget.m_events->m_events[DeviceType::Keyboard][EventType::Pressed]);
 		if(key_down_event)
 		{
 			if(m_key_down_handlers.find(key_down_event->m_code) != m_key_down_handlers.end())
 				m_key_down_handlers[key_down_event->m_code]();
 		}
 
-		KeyEvent* key_up_event = static_cast<KeyEvent*>(widget.m_events->m_events[size_t(DeviceType::Keyboard)][size_t(EventType::Released)]);
+		KeyEvent* key_up_event = static_cast<KeyEvent*>(widget.m_events->m_events[DeviceType::Keyboard][EventType::Released]);
 		if(key_up_event)
 		{
 			if(m_key_up_handlers.find(key_up_event->m_code) != m_key_up_handlers.end())
